pyproject.toml
[project]
name = "ai-lab-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.0.5",
    "duckduckgo-search>=7.2.0",
    "exa-py>=1.8.7",
    "google-generativeai>=0.8.4",
    "groq>=0.13.1",
    "lxml-html-clean>=0.4.1",
    "matplotlib>=3.10.0",
    "newspaper3k==0.2.8",
    "numpy>=2.2.1",
    "openai>=1.61.0",
    "pandas>=2.2.3",
    "pgvector>=0.3.6",
    "phidata>=2.7.6",
    "psycopg2>=2.9.10",
    "psycopg[binary]>=3.2.4",
    "pypdf>=5.3.0",
    "sqlalchemy>=2.0.38",
    "yfinance>=0.2.51",
    "youtube-transcript-api>=0.6.3",
]
